[TS-4] Rename: HomePage의 container 밑 태그들 Home.tsx로 이동
