+++ src/components/homepages/Home.style.ts
@@ -1,15 +1,12 @@
 import { css } from "@emotion/react";
 
-import { theme } from "@/styles/theme";
-
-// TODO: 추후 theme font & color로 변경
-export const homePageStyle = {
+export const homeStyle = {
 	container: css`
 		display: flex;
 		flex-direction: column;
 		min-height: 100vh;
 		min-width: 100%;
-		background-color: ${theme.color.bg};
+		background-color: white;
 	`,
 	innerWrapper: css`
 		display: flex;
